requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-ollama>=0.3.0

python-dotenv>=1.0.0
pydantic>=2.0
sqlalchemy>=2.0
fastapi>=0.100.0
uvicorn>=0.23.2
flask>=3.0.0